+++ cmake/toolchains/linux-x86_64-gcc.cmake
@@ -1,14 +1,30 @@
 # SPDX-FileCopyrightText: Copyright (c) 2022-2025 Oğuz Toraman <oguz.toraman@tutanota.com>
 # SPDX-License-Identifier: LGPL-3.0-only
 
-set(CMAKE_SYSTEM_NAME Linux)
-set(CMAKE_SYSTEM_PROCESSOR x86_64)
+set(CMAKE_SYSTEM_NAME
+    Linux
+)
 
-set(FILE_HOST x86_64-pc-linux-gnu)
+set(CMAKE_SYSTEM_PROCESSOR
+    x86_64
+)
 
-set(CMAKE_C_COMPILER gcc)
-set(CMAKE_CXX_COMPILER g++)
+set(FILE_HOST
+    x86_64-pc-linux-gnu
+)
 
-add_link_options(-fuse-ld=mold)
+set(CMAKE_C_COMPILER
+    gcc
+)
 
-add_compile_options(-Wall -Wextra -Wpedantic -Wfatal-errors)
+set(CMAKE_CXX_COMPILER
+    g++
+)
+
+add_link_options(
+    -fuse-ld=mold
+)
+
+add_compile_options(
+    -Wall -Wextra -Wpedantic -Wfatal-errors
+)