+++ cmake/toolchains/windows-x86_64-mingw64.cmake
@@ -1,14 +1,30 @@
 # SPDX-FileCopyrightText: Copyright (c) 2022-2025 Oğuz Toraman <oguz.toraman@tutanota.com>
 # SPDX-License-Identifier: LGPL-3.0-only
 
-set(CMAKE_SYSTEM_NAME Windows)
-set(CMAKE_SYSTEM_PROCESSOR x86_64)
+set(CMAKE_SYSTEM_NAME
+    Windows
+)
 
-set(FILE_HOST x86_64-pc-windows)
+set(CMAKE_SYSTEM_PROCESSOR
+    x86_64
+)
 
-set(CMAKE_C_COMPILER x86_64-w64-mingw32-gcc)
-set(CMAKE_CXX_COMPILER x86_64-w64-mingw32-g++)
+set(FILE_HOST
+    x86_64-pc-windows
+)
 
-add_compile_options(-Wall -Wextra -Wpedantic -Wfatal-errors)
+set(CMAKE_C_COMPILER
+    x86_64-w64-mingw32-gcc
+)
 
-link_libraries(stdc++exp)
+set(CMAKE_CXX_COMPILER
+    x86_64-w64-mingw32-g++
+)
+
+add_compile_options(
+    -Wall -Wextra -Wpedantic -Wfatal-errors
+)
+
+link_libraries(
+    stdc++exp
+)