+++ cmake/install.cmake
@@ -3,10 +3,41 @@
 
 install(
     FILES
+        ${magicxx_DEFAULT_DATABASE_FILES}
+    DESTINATION
+        ${magicxx_DEFAULT_DATABASES_INSTALL_DIR}
+    COMPONENT
+        Data
+)
+
+install(
+    FILES
+        ${magicxx_HEADER_FILES}
+    DESTINATION
+        ${CMAKE_INSTALL_INCLUDEDIR}/magicxx
+    COMPONENT
+        Development
+)
+
+install(
+    EXPORT
+        magicxxTargets
+    NAMESPACE
+        recognition::
+    DESTINATION
+        ${magicxx_CMAKE_INSTALL_LIBDIR}
+    COMPONENT
+        Development
+)
+
+install(
+    FILES
         ${magicxx_CMAKE_PACKAGE_CONFIG_FILE}
         ${magicxx_CMAKE_PACKAGE_CONFIG_VERSION_FILE}
     DESTINATION
         ${magicxx_CMAKE_INSTALL_LIBDIR}
+    COMPONENT
+        Development
 )
 
 install(
@@ -18,6 +49,8 @@ install(
         magicxxHeaders
     LIBRARY DESTINATION
         ${magicxx_INSTALL_LIBDIR}
+    COMPONENT
+        Library
 )
 
 install(
@@ -29,22 +62,8 @@ install(
         magicxx_staticHeaders
     ARCHIVE DESTINATION
         ${magicxx_INSTALL_LIBDIR}
-)
-
-install(
-    EXPORT
-        magicxxTargets
-    NAMESPACE
-        recognition::
-    DESTINATION
-        ${magicxx_CMAKE_INSTALL_LIBDIR}
-)
-
-install(
-    FILES
-        ${magicxx_DEFAULT_DATABASE_FILES}
-    DESTINATION
-        ${magicxx_DEFAULT_DATABASES_INSTALL_DIR}
+    COMPONENT
+        Archive
 )
 
 install(
@@ -52,6 +71,8 @@ install(
         documentation/html
     DESTINATION
         ${CMAKE_INSTALL_DOCDIR}
+    COMPONENT
+        Documentation
 )
 
 install(