+++ cmake/cpack_linux_rpm.cmake
@@ -1,0 +1,38 @@
+# SPDX-FileCopyrightText: Copyright (c) 2022-2025 Oğuz Toraman <oguz.toraman@tutanota.com>
+# SPDX-License-Identifier: LGPL-3.0-only
+
+set(CPACK_RPM_COMPONENT_INSTALL
+    OFF
+)
+
+set(CPACK_RPM_PACKAGE_EPOCH
+    1
+)
+
+set(CPACK_RPM_PACKAGE_RELEASE
+    1
+)
+
+set(CPACK_RPM_PACKAGE_ARCHITECTURE
+    ${CMAKE_SYSTEM_PROCESSOR}
+)
+
+set(CPACK_RPM_COMPRESSION_TYPE
+    zstd
+)
+
+set(CPACK_RPM_PACKAGE_LICENSE
+    "LGPL-3.0-only"
+)
+
+set(CPACK_RPM_PACKAGE_URL
+    ${magicxx_HOMEPAGE_URL}
+)
+
+set(CPACK_RPM_PACKAGE_AUTOREQPROV
+    YES
+)
+
+set(CPACK_RPM_CHANGELOG_FILE
+    ${magicxx_CHANGELOG_FILE}
+)