+++ cmake/files.cmake
@@ -5,6 +5,14 @@ set(magicxx_LICENSE_FILE
     ${magicxx_SOURCE_DIR}/COPYING.LESSER
 )
 
+set(magicxx_CHANGELOG_FILE
+    ${magicxx_SOURCE_DIR}/CHANGELOG.md
+)
+
+set(magicxx_CPACK_PACKAGE_DESCRIPTION_FILE
+    ${magicxx_SOURCE_DIR}/cmake/cpack_package_description_file.txt
+)
+
 set(magicxx_DEFAULT_DATABASE_FILE
     ${magicxx_DEFAULT_DATABASE_DIR}/magic
 )