+++ documentation/html/search/all_17.js
@@ -2,13 +2,13 @@ var searchData=
 [
   ['parameter_5fvalue_5fmap_5ft_0',['parameter_value_map_t',['../classrecognition_1_1magic.html#aa83f20ec1d8000f8f426d03a9a3eca9f',1,'recognition::magic']]],
   ['parameters_1',['parameters',['../classrecognition_1_1magic.html#a85e5da0093fe0bc74bb1911a99e1a940',1,'recognition::magic']]],
-  ['permanent_20ban_2',['4. Permanent Ban',['../md_CODE__OF__CONDUCT.html#autotoc_md45',1,'']]],
-  ['pledge_3',['Our Pledge',['../md_CODE__OF__CONDUCT.html#autotoc_md36',1,'']]],
+  ['permanent_20ban_2',['4. Permanent Ban',['../md_CODE__OF__CONDUCT.html#autotoc_md46',1,'']]],
+  ['pledge_3',['Our Pledge',['../md_CODE__OF__CONDUCT.html#autotoc_md37',1,'']]],
   ['policy_4',['Security Policy',['../md_SECURITY.html',1,'']]],
   ['preserve_5fatime_5',['preserve_atime',['../classrecognition_1_1magic.html#a41764bcc1857661422a1b062c9cdacdeac44aa990e942d6874811524fc0dbd8c0',1,'recognition::magic']]],
-  ['process_6',['Review Process',['../md_CONTRIBUTING.html#autotoc_md58',1,'']]],
-  ['project_20status_7',['Project Status',['../index.html#autotoc_md62',1,'']]],
-  ['project_20structure_8',['Project Structure',['../md_CONTRIBUTING.html#autotoc_md48',1,'']]],
-  ['project_20without_20installation_9',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md68',1,'']]],
-  ['pull_20request_10',['Creating a Pull Request',['../md_CONTRIBUTING.html#autotoc_md54',1,'']]]
+  ['process_6',['Review Process',['../md_CONTRIBUTING.html#autotoc_md59',1,'']]],
+  ['project_20status_7',['Project Status',['../index.html#autotoc_md63',1,'']]],
+  ['project_20structure_8',['Project Structure',['../md_CONTRIBUTING.html#autotoc_md49',1,'']]],
+  ['project_20without_20installation_9',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md69',1,'']]],
+  ['pull_20request_10',['Creating a Pull Request',['../md_CONTRIBUTING.html#autotoc_md55',1,'']]]
 ];