+++ documentation/html/search/all_18.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['quality_0',['Code Quality',['../index.html#autotoc_md63',1,'']]]
+  ['quality_0',['Code Quality',['../index.html#autotoc_md64',1,'']]]
 ];