+++ documentation/html/search/all_1c.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['uninstalling_20libmagicxx_0',['Uninstalling Libmagicxx',['../index.html#autotoc_md67',1,'']]],
-  ['using_20installed_20libmagicxx_20with_20cmake_1',['Using Installed Libmagicxx with CMake',['../index.html#autotoc_md69',1,'']]],
-  ['using_20libmagicxx_20in_20a_20cmake_20project_20without_20installation_2',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md68',1,'']]],
+  ['uninstalling_20libmagicxx_0',['Uninstalling Libmagicxx',['../index.html#autotoc_md68',1,'']]],
+  ['using_20installed_20libmagicxx_20with_20cmake_1',['Using Installed Libmagicxx with CMake',['../index.html#autotoc_md70',1,'']]],
+  ['using_20libmagicxx_20in_20a_20cmake_20project_20without_20installation_2',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md69',1,'']]],
   ['utility_3',['utility',['../namespaceutility.html',1,'']]],
   ['utility_2ehpp_4',['utility.hpp',['../utility_8hpp.html',1,'']]],
   ['utility_3a_3arange_5fcontainer_5',['range_container',['../conceptutility_1_1range__container.html',1,'utility']]],