+++ documentation/html/search/all_1e.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['warning_0',['2. Warning',['../md_CODE__OF__CONDUCT.html#autotoc_md43',1,'']]],
-  ['with_20cmake_1',['Using Installed Libmagicxx with CMake',['../index.html#autotoc_md69',1,'']]],
-  ['without_20installation_2',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md68',1,'']]]
+  ['warning_0',['2. Warning',['../md_CODE__OF__CONDUCT.html#autotoc_md44',1,'']]],
+  ['with_20cmake_1',['Using Installed Libmagicxx with CMake',['../index.html#autotoc_md70',1,'']]],
+  ['without_20installation_2',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md69',1,'']]]
 ];