+++ documentation/html/search/all_9.js
@@ -1,9 +1,9 @@
 var searchData=
 [
-  ['a_20cmake_20project_20without_20installation_0',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md68',1,'']]],
-  ['a_20pull_20request_1',['Creating a Pull Request',['../md_CONTRIBUTING.html#autotoc_md54',1,'']]],
-  ['a_20vulnerability_2',['Reporting a Vulnerability',['../md_SECURITY.html#autotoc_md80',1,'']]],
-  ['an_20issue_3',['How to Create an Issue',['../md_CONTRIBUTING.html#autotoc_md56',1,'']]],
+  ['a_20cmake_20project_20without_20installation_0',['Using Libmagicxx in a CMake Project Without Installation',['../index.html#autotoc_md69',1,'']]],
+  ['a_20pull_20request_1',['Creating a Pull Request',['../md_CONTRIBUTING.html#autotoc_md55',1,'']]],
+  ['a_20vulnerability_2',['Reporting a Vulnerability',['../md_SECURITY.html#autotoc_md81',1,'']]],
+  ['an_20issue_3',['How to Create an Issue',['../md_CONTRIBUTING.html#autotoc_md57',1,'']]],
   ['apple_4',['apple',['../classrecognition_1_1magic.html#a41764bcc1857661422a1b062c9cdacdeaa9741451e5183c02e81018e51b7b7a10',1,'recognition::magic']]],
-  ['attribution_5',['Attribution',['../md_CODE__OF__CONDUCT.html#autotoc_md46',1,'']]]
+  ['attribution_5',['Attribution',['../md_CODE__OF__CONDUCT.html#autotoc_md47',1,'']]]
 ];